# dashboard/openapi.json
{
  "components": {
    "schemas": {
      "ConsolidateBody": {
        "properties": {
          "min_cluster_size": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Cluster Size"
          },
          "n_clusters": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "N Clusters"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          }
        },
        "title": "ConsolidateBody",
        "type": "object"
      },
      "ConsolidateResult": {
        "properties": {
          "schemas": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Schemas",
            "type": "array"
          },
          "schemas_formed": {
            "title": "Schemas Formed",
            "type": "integer"
          }
        },
        "required": [
          "schemas_formed",
          "schemas"
        ],
        "title": "ConsolidateResult",
        "type": "object"
      },
...
      "DeleteResult": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "deleted_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Deleted Ids",
            "type": "array"
          }
        },
        "required": [
          "deleted_ids",
          "count"
        ],
        "title": "DeleteResult",
        "type": "object"
      },
...
      "ForgetBody": {
        "properties": {
          "decay_tau_seconds": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decay Tau Seconds"
          },
          "dry_run": {
            "default": true,
            "title": "Dry Run",
            "type": "boolean"
          },
          "max_scan": {
            "default": 1000,
            "maximum": 100000.0,
            "minimum": 1.0,
            "title": "Max Scan",
            "type": "integer"
          },
          "strength_floor": {
            "default": 0.05,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Strength Floor",
            "type": "number"
          }
        },
        "title": "ForgetBody",
        "type": "object"
      },
      "ForgetResult": {
        "properties": {
          "candidate_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Candidate Ids",
            "type": "array"
          },
          "cold_schema_candidate_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Cold Schema Candidate Ids",
            "type": "array"
          },
          "dry_run": {
            "title": "Dry Run",
            "type": "boolean"
          },
          "removed": {
            "title": "Removed",
            "type": "integer"
          },
          "scanned": {
            "title": "Scanned",
            "type": "integer"
          },
          "schemas_removed": {
            "title": "Schemas Removed",
            "type": "integer"
          },
          "schemas_scanned": {
            "title": "Schemas Scanned",
            "type": "integer"
          }
        },
        "required": [
          "candidate_ids",
          "removed",
          "dry_run",
          "scanned",
          "schemas_scanned",
          "cold_schema_candidate_ids",
          "schemas_removed"
        ],
        "title": "ForgetResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MemoryBrowse": {
        "description": "``GET /memories`` without ``q`` \u2014 paginated raw listing.",
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/MemoryListItem"
            },
            "title": "Items",
            "type": "array"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          }
        },
        "required": [
          "kind",
          "items",
          "count",
          "offset"
        ],
        "title": "MemoryBrowse",
        "type": "object"
      },
      "MemoryDetailResponse": {
        "properties": {
          "group_id": {
            "title": "Group Id",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "plasticity_edges": {
            "items": {
              "$ref": "#/components/schemas/PlasticityEdgeDetail"
            },
            "title": "Plasticity Edges",
            "type": "array"
          },
          "segments": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Segments",
            "type": "array"
          },
          "sr_transitions": {
            "items": {
              "$ref": "#/components/schemas/SrTransition"
            },
            "title": "Sr Transitions",
            "type": "array"
          },
          "sr_transitions_in": {
            "items": {
              "$ref": "#/components/schemas/SrTransitionIn"
            },
            "title": "Sr Transitions In",
            "type": "array"
          }
        },
        "required": [
          "id",
          "group_id",
          "segments",
          "sr_transitions",
          "sr_transitions_in",
          "plasticity_edges"
        ],
        "title": "MemoryDetailResponse",
        "type": "object"
      },
      "MemoryListItem": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          }
        },
        "required": [
          "id",
          "content",
          "metadata"
        ],
        "title": "MemoryListItem",
        "type": "object"
      },
      "MemorySearch": {
        "description": "``GET /memories?q=`` \u2014 hybrid recall + substring search.\n\nHit items carry open keys (distance, score, source, metadata,\nsubstring_offset) so they stay ``dict[str, Any]`` \u2014 modelling them\nstrictly would drop fields the UI relies on.",
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Items",
            "type": "array"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "recall_mode": {
            "enum": [
              "semantic",
              "substring",
              "hybrid",
              "empty"
            ],
            "title": "Recall Mode",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "query",
          "items",
          "count",
          "recall_mode"
        ],
        "title": "MemorySearch",
        "type": "object"
      },
...
      "PlasticityEdgeDetail": {
        "properties": {
          "bonus": {
            "title": "Bonus",
            "type": "number"
          },
          "dst": {
            "title": "Dst",
            "type": "integer"
          },
          "hits": {
            "title": "Hits",
            "type": "integer"
          },
          "src": {
            "title": "Src",
            "type": "integer"
          },
          "weight": {
            "title": "Weight",
            "type": "number"
          }
        },
        "required": [
          "src",
          "dst",
          "weight",
          "bonus",
          "hits"
        ],
        "title": "PlasticityEdgeDetail",
        "type": "object"
      },
      "ReflectBody": {
        "properties": {
          "k": {
            "default": 5,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "K",
            "type": "integer"
          },
          "query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id"
        ],
        "title": "ReflectBody",
        "type": "object"
      },
      "SemanticDetailResponse": {
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "source_episode_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Source Episode Ids",
            "type": "array"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "number"
          },
          "user_asserted": {
            "title": "User Asserted",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "content",
          "kind",
          "tags",
          "confidence",
          "user_asserted",
          "source_episode_ids",
          "created_at",
          "updated_at"
        ],
        "title": "SemanticDetailResponse",
        "type": "object"
      },
...
      "SrTransition": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "dst": {
            "title": "Dst",
            "type": "integer"
          }
        },
        "required": [
          "dst",
          "count"
        ],
        "title": "SrTransition",
        "type": "object"
      },
      "SrTransitionIn": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "src": {
            "title": "Src",
            "type": "integer"
          }
        },
        "required": [
          "src",
          "count"
        ],
        "title": "SrTransitionIn",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Synara Dashboard",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/admin/consolidate": {
      "post": {
        "operationId": "admin_consolidate_api_admin_consolidate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsolidateBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsolidateResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Consolidate",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/forget": {
      "post": {
        "operationId": "admin_forget_api_admin_forget_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgetBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForgetResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Forget",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/reflect": {
      "post": {
        "operationId": "admin_reflect_api_admin_reflect_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReflectBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Admin Reflect Api Admin Reflect Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Reflect",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/memories": {
      "get": {
        "operationId": "list_memories_api_memories_get",
        "parameters": [
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "default": "episodic",
              "enum": [
                "episodic",
                "semantic"
              ],
              "title": "Kind",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 8000,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 100000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/MemoryBrowse"
                    },
                    {
                      "$ref": "#/components/schemas/MemorySearch"
                    }
                  ],
                  "title": "Response List Memories Api Memories Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Memories",
        "tags": [
          "memories"
        ]
      }
    },
    "/api/memories/{episode_id}": {
      "delete": {
        "operationId": "delete_memory_api_memories__episode_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "episode_id",
            "required": true,
            "schema": {
              "minimum": 0,
              "title": "Episode Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Memory",
        "tags": [
          "memories"
        ]
      },
      "get": {
        "operationId": "memory_detail_api_memories__episode_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "episode_id",
            "required": true,
            "schema": {
              "minimum": 0,
              "title": "Episode Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Memory Detail",
        "tags": [
          "memories"
        ]
      }
    },
...
    "/api/semantic/{semantic_id}": {
      "get": {
        "operationId": "semantic_detail_api_semantic__semantic_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "semantic_id",
            "required": true,
            "schema": {
              "minimum": 0,
              "title": "Semantic Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SemanticDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Semantic Detail",
        "tags": [
          "memories"
        ]
      }
    },
...
  }
}